+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - [browser] fix: DOMError and DOMException should be error level events
+- [utils] fix: Dont mutate original input in decycle util function
 - [utils] ref: Update wrap method to hide internal sentry flags
 - [utils] fix: Make internal Sentry flags non-enumerable in fill util